+++ src/pyramid_bimt/tests/test_auditlog_model.py
@@ -31,8 +31,8 @@ class TestAuditLogEventType(unittest.TestCase):
 
     def test__repr__(self):
         self.assertEqual(
-            repr(AuditLogEventType.by_id(1)),
-            '<AuditLogEventType:1 (name=u\'SanityCheckDone\')>'
+            repr(AuditLogEventType.by_name('SanityCheckDone')),
+            '<AuditLogEventType:2 (name=u\'SanityCheckDone\')>'
         )
 
     def test_default_event_types(self):
@@ -40,44 +40,48 @@ class TestAuditLogEventType(unittest.TestCase):
         from pyramid_bimt.scripts.populate import default_audit_log_event_types
         types = default_audit_log_event_types()
 
-        self.assertEqual(types[0].name, 'SanityCheckDone')
-        self.assertEqual(types[0].title, 'Sanity Check Done')
-        self.assertEqual(types[0].description, 'Emitted whenever a sanity check is done.')  # noqa
+        self.assertEqual(types[0].name, 'ReferralEmailSent')
+        self.assertEqual(types[0].title, 'Referral Email Sent')
+        self.assertEqual(types[0].description, 'Emitted whenever a user sends a referral email.')  # noqa
 
-        self.assertEqual(types[1].name, 'UserChangedPassword')
-        self.assertEqual(types[1].title, 'User Changed Password')
-        self.assertEqual(types[1].description, 'Emitted whenever a user changes its password.')  # noqa
+        self.assertEqual(types[1].name, 'SanityCheckDone')
+        self.assertEqual(types[1].title, 'Sanity Check Done')
+        self.assertEqual(types[1].description, 'Emitted whenever a sanity check is done.')  # noqa
 
-        self.assertEqual(types[2].name, 'UserCreated')
-        self.assertEqual(types[2].title, 'User Created')
-        self.assertEqual(types[2].description, 'Emitted whenever a new user is created.')  # noqa
+        self.assertEqual(types[2].name, 'UserChangedPassword')
+        self.assertEqual(types[2].title, 'User Changed Password')
+        self.assertEqual(types[2].description, 'Emitted whenever a user changes its password.')  # noqa
 
-        self.assertEqual(types[3].name, 'UserDisabled')
-        self.assertEqual(types[3].title, 'User Disabled')
-        self.assertEqual(types[3].description, 'Emitted whenever a user is disabled.')  # noqa
+        self.assertEqual(types[3].name, 'UserCreated')
+        self.assertEqual(types[3].title, 'User Created')
+        self.assertEqual(types[3].description, 'Emitted whenever a new user is created.')  # noqa
 
-        self.assertEqual(types[4].name, 'UserEnabled')
-        self.assertEqual(types[4].title, 'User Enabled')
-        self.assertEqual(types[4].description, 'Emitted whenever a user is enabled.')  # noqa
+        self.assertEqual(types[4].name, 'UserDisabled')
+        self.assertEqual(types[4].title, 'User Disabled')
+        self.assertEqual(types[4].description, 'Emitted whenever a user is disabled.')  # noqa
 
-        self.assertEqual(types[5].name, 'UserLoggedIn')
-        self.assertEqual(types[5].title, 'User Logged In')
-        self.assertEqual(types[5].description, 'Emitted whenever a user logs in.')  # noqa
+        self.assertEqual(types[5].name, 'UserEnabled')
+        self.assertEqual(types[5].title, 'User Enabled')
+        self.assertEqual(types[5].description, 'Emitted whenever a user is enabled.')  # noqa
 
-        self.assertEqual(types[6].name, 'UserLoggedInAs')
-        self.assertEqual(types[6].title, 'User Logged In As')
-        self.assertEqual(types[6].description, 'Emitted whenever a user logs in as another user.')  # noqa
+        self.assertEqual(types[6].name, 'UserLoggedIn')
+        self.assertEqual(types[6].title, 'User Logged In')
+        self.assertEqual(types[6].description, 'Emitted whenever a user logs in.')  # noqa
 
-        self.assertEqual(types[7].name, 'UserLoggedOut')
-        self.assertEqual(types[7].title, 'User Logged Out')
-        self.assertEqual(types[7].description, 'Emitted whenever a user logs out.')  # noqa
+        self.assertEqual(types[7].name, 'UserLoggedInAs')
+        self.assertEqual(types[7].title, 'User Logged In As')
+        self.assertEqual(types[7].description, 'Emitted whenever a user logs in as another user.')  # noqa
+
+        self.assertEqual(types[8].name, 'UserLoggedOut')
+        self.assertEqual(types[8].title, 'User Logged Out')
+        self.assertEqual(types[8].description, 'Emitted whenever a user logs out.')  # noqa
 
     def test_get_event_type_by_id(self):
         from pyramid_bimt.models import AuditLogEventType
 
-        self.assertEqual(AuditLogEventType.by_id(2).name, 'UserChangedPassword')  # noqa
-        self.assertEqual(AuditLogEventType.by_id(2).title, 'User Changed Password')  # noqa
-        self.assertEqual(AuditLogEventType.by_id(2).description, 'Emitted whenever a user changes its password.')  # noqa
+        self.assertEqual(AuditLogEventType.by_id(3).name, 'UserChangedPassword')  # noqa
+        self.assertEqual(AuditLogEventType.by_id(3).title, 'User Changed Password')  # noqa
+        self.assertEqual(AuditLogEventType.by_id(3).description, 'Emitted whenever a user changes its password.')  # noqa
 
     def test_get_event_type_get_all(self):
         from pyramid_bimt.scripts.populate import default_audit_log_event_types
@@ -95,7 +99,7 @@ class TestAuditLogEventType(unittest.TestCase):
     def test_get_event_type_by_name(self):
         from pyramid_bimt.models import AuditLogEventType
 
-        self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').id, 2)  # noqa
+        self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').id, 3)  # noqa
         self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').title, 'User Changed Password')  # noqa
         self.assertEqual(AuditLogEventType.by_name('UserChangedPassword').description, 'Emitted whenever a user changes its password.')  # noqa
 
@@ -136,7 +140,7 @@ class TestAuditLogEntryModel(unittest.TestCase):
         )
         self.assertEqual(
             repr(entry),
-            '<AuditLogEntry:1 (user=\'foo@bar.com\', type=u\'SanityCheckDone\')>'  # noqa
+            '<AuditLogEntry:1 (user=\'foo@bar.com\', type=u\'ReferralEmailSent\')>'  # noqa
         )
 
     def test__repr__empty_values(self):